Check file given as first argument
Check the file whose name has been specified on the command line for
errors, instead of `stdin'. Git passes the file name of the file
containing the commit message instead of passing the commit message to
`stdin'.

This was supposed to be committed in the first commit. but somehow it
got lost.