Add -h and -0 command-line options
Add the `-0' option so that a flycheck checker can be written. These
don't seem to like it when the checker exits with a non-zero status,
whereas for the purposes of being a `commit-msg' hook for git it is
important that it does exit with a non-zero status to stop git from
committing.